Refactor & source layout:
- Rename the project/group ID from `com.github.miloszpp` to `org.scala-ts`
- Split project in two modules/artifacts `"org.scala-ts" %% "scala-ts-core"` and `"org.scala-ts" % "scala-ts-sbt"`
- Base package `com.mpc.scalats` is renamed to `org.scalats`
