+++ core/src/main/scala-2.10/org/scalats/core/Internals.scala
@@ -1,0 +1,8 @@
+package org.scalats.core
+
+import scala.collection.immutable.ListSet
+
+private[core] object Internals {
+  /** With predictable order (for emitter). */
+  def list[T](set: ListSet[T]): List[T] = set.toList.reverse
+}
