+++ core/src/main/scala/org/scalats/core/StableValue.java
@@ -1,0 +1,9 @@
+package org.scalats.core;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface StableValue {
+    String typescript();
+}
