--- src/main/scala-2.12/com/mpc/scalats/core/Internals.scala
@@ -1,8 +1,0 @@
-package com.mpc.scalats.core
-
-import scala.collection.immutable.ListSet
-
-private[core] object Internals {
-  /** With predictable order (for emitter). */
-  def list[T](set: ListSet[T]): List[T] = set.toList
-}
