--- src/main/scala/com/mpc/scalats/core/StableValue.java
@@ -1,9 +1,0 @@
-package com.mpc.scalats.core;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface StableValue {
-    String typescript();
-}
